requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1
sarvamai>=0.1.0
pydub==0.25.1
pandas==2.2.2
aiofiles==23.2.1
